# twitter-swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Twitter-Clone
  description: |-
    This is a sample Twitter Clone Server based on the OpenAPI 3.0 specification. You can explore and interact with the API to understand its capabilities and help us enhance it further.
    For more information, visit [https://github.com/phannguyenminhphat1/Nodejsts.git](https://github.com/phannguyenminhphat1/Nodejsts.git). We are continuously working to improve the API and add new functionalities. contact:
    email: phannguyenminhphat1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://minhphat.io.vn
...
paths:
...
  /api/users/refresh-token:
    post:
      tags:
        - Users
      description: Refresh token
      operationId: refresh_token
      security:
        - BearerAuth: []
      requestBody:
        description: Refresh token
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: eyblksakmOmasQsnncjsjnsmwoa5wx7tYNS
        required: true
      responses:
        '200':
          description: Refresh token successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token successfully
                  result:
                    $ref: '#/components/schemas/SuccessNewAuthentication'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Used refresh token or not exist
        '422':
          description: Validation error

  # /api/users/email_verify_token
  /api/users/email_verify_token:
    post:
      tags:
        - Users
      description: Email verify token
      operationId: email_verify_token
      requestBody:
        description: Email verify token
        content:
          application/json:
            schema:
              type: object
              properties:
                email_verify_token:
                  type: string
                  example: eyblksakmOmasQsnncjsjnsmwoa5wx7tYNS
        required: true
      responses:
        '200':
          description: Verify email successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email verify successfully
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid signature
        '422':
          description: Validation error

  # /api/users/resend-verify-email
...
  /api/users/twitter_circle:
    post:
      tags:
        - Users
      description: Add user in twitter circle
      operationId: twitter_circle
      security:
        - BearerAuth: []
      requestBody:
        description: Add user twitter circle by username
        content:
          application/json:
            schema:
              type: object
              properties:
                twitter_circle:
                  type: array
                  items:
                    type: string
                  example:
                    - freshernodejs
                    - username
        required: true
      responses:
        '200':
          description: Add twitter circle successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Add twitter circle successfully
        '401':
          description: Unauthorized
        '422':
          description: Validation error

  # /api/users/unfollow
...
components:
  schemas:
...
    SuccessNewAuthentication:
      type: object
      properties:
        new_access_token:
          type: string
          example: eyblksakmOmasQsnncjsjnsmwoa5wx7tYNS
        new_refresh_token:
          type: string
          example: eyblksakmOmasQsnncjsjnsmwoa5wx7tYNS
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT